fix: address PR review issues for hud-hzub.2
- renderer.rs: clamp UTF-8 slice offset to a valid char boundary via
  is_char_boundary() walk-back, preventing panic on non-boundary offsets
- renderer.rs: update StreamRevealState doc comment — breakpoints are
  expected non-decreasing but not enforced here; compositor is safe either way
- types.rs: change ZonePublishRecord.breakpoints from Vec<usize> to Vec<u64>
  for stable wire format across 32/64-bit platforms
- graph.rs: update publish_to_zone_with_breakpoints and
  publish_to_zone_with_lease_and_breakpoints signatures to Vec<u64>
- renderer.rs: StreamRevealState::new converts Vec<u64> -> Vec<usize> internally
- tools.rs: validate that breakpoints are only sent with StreamText content;
  return invalid_params error otherwise
- tools.rs: change PublishToZoneParams.breakpoints to Vec<u64> to match record type

+++ crates/tze_hud_scene/src/graph.rs
@@ -2656,7 +2656,7 @@ impl SceneGraph {
         content: ZoneContent,
         publisher_namespace: &str,
         merge_key: Option<String>,
-        breakpoints: Vec<usize>,
+        breakpoints: Vec<u64>,
     ) -> Result<(), ValidationError> {
         use crate::lease::orphan::ZonePublishResult;
 
@@ -2740,7 +2740,7 @@ impl SceneGraph {
         merge_key: Option<String>,
         expires_at_wall_us: Option<u64>,
         content_classification: Option<String>,
-        breakpoints: Vec<usize>,
+        breakpoints: Vec<u64>,
     ) -> Result<(), ValidationError> {
         // Check zone exists and content type is accepted
         let (contention_policy, max_publishers, accepted) = {
